Further consolidate storage service fixtures
